fix: ID 못 찾을 경우 error 메세지 수정

+++ src/services/articleService.js
@@ -1,3 +1,4 @@
+import NotFoundError from "../lib/error/NotFoundError.js";
 import articleRepository from "../repositories/articleRepository.js";
 
 async function getAll({ offset, limit, order, search }) {
@@ -24,15 +25,30 @@ async function create(article) {
   return articleRepository.save(article);
 }
 
-async function getId(id) {
-  return articleRepository.getId(id);
+async function getById(id) {
+  const article = await articleRepository.getById(id);
+
+  if (!article) {
+    throw new NotFoundError(id);
+  }
+  return article;
 }
 
 async function update(id, article) {
-  return articleRepository.update(id, article);
+  const findArticle = await articleRepository.getById(id);
+  if (!findArticle) {
+    throw new NotFoundError(id);
+  }
+  return await articleRepository.update(id, article);
 }
 
 async function deleteById(id) {
-  return articleRepository.deleteById(id);
+  const findArticle = await articleRepository.getById(id);
+
+  if (!findArticle) {
+    throw new NotFoundError(id);
+  }
+
+  return await articleRepository.deleteById(id);
 }
-export default { getAll, create, getId, update, deleteById };
+export default { getAll, create, getById, update, deleteById };
