+++ src/services/productService.js
@@ -1,3 +1,4 @@
+import NotFoundError from "../lib/error/NotFoundError.js";
 import productRepository from "../repositories/productRepository.js";
 
 async function getAll({ offset, limit, order, search }) {
@@ -25,15 +26,31 @@ async function createProduct(product) {
 }
 
 async function getById(id) {
-  return productRepository.getById(id);
+  const product = await productRepository.getById(id);
+
+  if (!product) {
+    throw new NotFoundError(id);
+  }
+
+  return product;
 }
 
 async function update(id, product) {
-  return productRepository.update(id, product);
+  const findProduct = await productRepository.getById(id);
+  if (!findProduct) {
+    throw new NotFoundError(id);
+  }
+
+  return await productRepository.update(id, product);
 }
 
 async function deleteProduct(id) {
-  return productRepository.deleteProduct(id);
+  const findProduct = await productRepository.getById(id);
+  if (!findProduct) {
+    throw new NotFoundError(id);
+  }
+
+  return await productRepository.deleteProduct(id);
 }
 
 export default { getAll, createProduct, getById, update, deleteProduct };
